requirements.txt
discord.py==0.14.3
aiohttp
websockets
chardet
beautifulsoup4